Remove direct package dependency on DD
