+++ Geometry/EcalCommonData/plugins/BuildFile.xml
@@ -1,4 +1,5 @@
-<library   file="module.cc" name="GeometryEcalCommonDataPlugins">
+<library   file="*.cc" name="GeometryEcalCommonDataPlugins">
+  <use   name="DetectorDescription/Core"/>
   <use   name="Geometry/EcalCommonData"/>
 </library>
 <flags   EDM_PLUGIN="1"/>
